Migrate to OAuth2 and implement Placement Selection

+++ src/main/java/com/iiitb/erp/placement/entity/Organisation.java
@@ -10,7 +10,7 @@ public class Organisation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id; // Changed from Long to Integer
 
     private String name;
