+++ src/main/java/com/iiitb/erp/placement/repository/EmployeeRepository.java
@@ -2,10 +2,8 @@ package com.iiitb.erp.placement.repository;
 
 import com.iiitb.erp.placement.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
     Optional<Employee> findByEmail(String email);
 }
